Don't specify a default port in the scp command
The convenient way to use a nonstandard ssh port is to specify a value
in ~/.ssh/config.  By appending the default port to the scp command,
scpaste overrides the port value a user may have set in ~/.ssh/config.
With this patch, scpaste only supplies a port in the scp command if the
user has set the variable `scpaste-scp-port'.